move login/signup to navbar
